fix: clang warnings on C code
